+++ lib/miniswen/local.rb
@@ -27,8 +27,6 @@ module Miniswen
       end
     end
 
-    def stop = nil
-
     private
 
     def spawn_arguments(command, env) = [ env || {}, "sh", "-c", command ]